+++ client/src/components/MainScreen.js
@@ -1,0 +1,24 @@
+import ToBtn from './ToBtn'
+
+const APP_STATUS = {
+  USER_NOT_LOGGED: 'user_not_logged',
+  USER_TO_SIGNIN: 'user_to_signin',
+  USER_TO_REGISTER: 'user_to_register',
+  USER_LOGGED: 'user_logged',
+}
+
+export default function MainScreen({ setStatus }) {
+  const handlerToLogin = () => setStatus(APP_STATUS.USER_TO_SIGNIN)
+  const handlerToRegister = () => setStatus(APP_STATUS.USER_TO_REGISTER)
+
+  return (
+    <>
+      <h2>Welcome to ChatJS</h2>
+      <p>To use the chat, first signup or login with 'Anonymous' to try it!</p>
+      <div className='button_container'>
+        <ToBtn onClick={handlerToLogin}>to Login</ToBtn>
+        <ToBtn onClick={handlerToRegister}>to Registry</ToBtn>
+      </div>
+    </>
+  )
+}
