+++ client/src/components/ToBtn.js
@@ -1,0 +1,7 @@
+export default function ToBtn({ onClick, children }) {
+  return (
+    <button className='register_btn' onClick={() => onClick()}>
+      {children}
+    </button>
+  )
+}
